+++ blender_asset_tracer/blendfile/__init__.py
@@ -49,7 +49,7 @@ def open_cached(path: pathlib.Path, mode='rb',
                 assert_cached: typing.Optional[bool] = None) -> 'BlendFile':
     """Open a blend file, ensuring it is only opened once."""
     my_log = log.getChild('open_cached')
-    bfile_path = path.absolute().resolve()
+    bfile_path = bpathlib.make_absolute(path)
 
     if assert_cached is not None:
         is_cached = bfile_path in _cached_bfiles
@@ -84,13 +84,13 @@ def close_all_cached() -> None:
 
 def _cache(path: pathlib.Path, bfile: 'BlendFile'):
     """Add a BlendFile to the cache."""
-    bfile_path = path.absolute().resolve()
+    bfile_path = bpathlib.make_absolute(path)
     _cached_bfiles[bfile_path] = bfile
 
 
 def _uncache(path: pathlib.Path):
     """Remove a BlendFile object from the cache."""
-    bfile_path = path.absolute().resolve()
+    bfile_path = bpathlib.make_absolute(path)
     _cached_bfiles.pop(bfile_path, None)