+++ blender_asset_tracer/bpathlib.py
@@ -25,6 +25,7 @@ or vice versa.
 
 import os.path
 import pathlib
+import platform
 import string
 import sys
 
@@ -41,18 +42,29 @@ class BlendPath(bytes):
         return super().__new__(cls, path.replace(b'\\', b'/'))
 
     @classmethod
-    def mkrelative(cls, asset_path: pathlib.Path, bfile_path: pathlib.PurePath) -> 'BlendPath':
+    def mkrelative(cls, asset_path: pathlib.PurePath, bfile_path: pathlib.PurePath) -> 'BlendPath':
         """Construct a BlendPath to the asset relative to the blend file.
 
         Assumes that bfile_path is absolute.
+
+        Note that this can return an absolute path on Windows when 'asset_path'
+        and 'bfile_path' are on different drives.
         """
         from collections import deque
 
+        # Only compare absolute paths.
         assert bfile_path.is_absolute(), \
             'BlendPath().mkrelative(bfile_path=%r) should get absolute bfile_path' % bfile_path
+        assert asset_path.is_absolute(), \
+            'BlendPath().mkrelative(asset_path=%r) should get absolute asset_path' % asset_path
+
+        # There is no way to construct a relative path between drives.
+        if bfile_path.drive != asset_path.drive:
+            return cls(asset_path)
 
         bdir_parts = deque(bfile_path.parent.parts)
-        asset_parts = deque(asset_path.absolute().parts)
+        asset_path = make_absolute(asset_path)
+        asset_parts = deque(asset_path.parts)
 
         # Remove matching initial parts. What is left in bdir_parts represents
         # the number of '..' we need. What is left in asset_parts represents
@@ -101,6 +113,8 @@ class BlendPath(bytes):
 
         Interprets the path as UTF-8, and if that fails falls back to the local
         filesystem encoding.
+
+        The exact type returned is determined by the current platform.
         """
         # TODO(Sybren): once we target Python 3.6, implement __fspath__().
         try:
@@ -146,3 +160,53 @@ class BlendPath(bytes):
         else:
             my_relpath = self
         return BlendPath(os.path.join(root, my_relpath))
+
+
+def make_absolute(path: pathlib.PurePath) -> pathlib.Path:
+    """Make the path absolute without resolving symlinks or drive letters.
+
+    This function is an alternative to `Path.resolve()`. It make the path absolute,
+    and resolves `../../`, but contrary to `Path.resolve()` does NOT perform these
+    changes:
+        - Symlinks are NOT followed.
+        - Windows Network shares that are mapped to a drive letter are NOT resolved
+          to their UNC notation.
+
+    The type of the returned path is determined by the current platform.
+    """
+    str_path = path.as_posix()
+    if len(str_path) >= 2 and str_path[0].isalpha() and str_path[1] == ':':
+        # This is an absolute Windows path. It must be handled with care on non-Windows platforms.
+        if platform.system() != 'Windows':
+            # Normalize the POSIX-like part of the path, but leave out the drive letter.
+            non_drive_path = str_path[2:]
+            normalized = os.path.normpath(non_drive_path)
+            # Stick the drive letter back on the normalized path.
+            return pathlib.Path(str_path[:2] + normalized)
+
+    return pathlib.Path(os.path.abspath(str_path))
+
+
+def strip_root(path: pathlib.PurePath) -> pathlib.PurePosixPath:
+    """Turn the path into a relative path by stripping the root.
+
+    This also turns any drive letter into a normal path component.
+
+    This changes "C:/Program Files/Blender" to "C/Program Files/Blender",
+    and "/absolute/path.txt" to "absolute/path.txt", making it possible to
+    treat it as a relative path.
+    """
+
+    if path.drive:
+        return pathlib.PurePosixPath(path.drive[0], *path.parts[1:])
+    if isinstance(path, pathlib.PurePosixPath):
+        # This happens when running on POSIX but still handling paths
+        # originating from a Windows machine.
+        parts = path.parts
+        if parts and len(parts[0]) == 2 and parts[0][0].isalpha() and parts[0][1] == ':':
+            # The first part is a drive letter.
+            return pathlib.PurePosixPath(parts[0][0], *path.parts[1:])
+
+    if path.is_absolute():
+        return pathlib.PurePosixPath(*path.parts[1:])
+    return pathlib.PurePosixPath(path)