+++ tests/test_bpathlib.py
@@ -1,8 +1,11 @@
-from pathlib import Path, PurePosixPath
+import os
+from pathlib import Path, PurePath, PurePosixPath, PureWindowsPath
+import platform
+import tempfile
 import unittest
 from unittest import mock
 
-from blender_asset_tracer.bpathlib import BlendPath
+from blender_asset_tracer.bpathlib import BlendPath, make_absolute, strip_root
 
 
 class BlendPathTest(unittest.TestCase):
@@ -29,11 +32,11 @@ class BlendPathTest(unittest.TestCase):
         self.assertEqual("BlendPath(b'//some/file.blend')", repr(p))
 
     def test_to_path(self):
-        self.assertEqual(PurePosixPath('/some/file.blend'),
+        self.assertEqual(PurePath('/some/file.blend'),
                          BlendPath(b'/some/file.blend').to_path())
-        self.assertEqual(PurePosixPath('C:/some/file.blend'),
+        self.assertEqual(PurePath('C:/some/file.blend'),
                          BlendPath(b'C:/some/file.blend').to_path())
-        self.assertEqual(PurePosixPath('C:/some/file.blend'),
+        self.assertEqual(PurePath('C:/some/file.blend'),
                          BlendPath(br'C:\some\file.blend').to_path())
 
         with mock.patch('sys.getfilesystemencoding') as mock_getfse:
@@ -41,7 +44,7 @@ class BlendPathTest(unittest.TestCase):
 
             # \xe9 is Latin-1 for é, and BlendPath should revert to using the
             # (mocked) filesystem encoding when decoding as UTF-8 fails.
-            self.assertEqual(PurePosixPath('C:/some/filé.blend'),
+            self.assertEqual(PurePath('C:/some/filé.blend'),
                              BlendPath(b'C:\\some\\fil\xe9.blend').to_path())
 
         with self.assertRaises(ValueError):
@@ -84,7 +87,8 @@ class BlendPathTest(unittest.TestCase):
         self.assertEqual(BlendPath(b'//root/parent.blend'),
                          BlendPath(b'//root/') / b'parent.blend')
 
-    def test_mkrelative(self):
+    @unittest.skipIf(platform.system() == 'Windows', "POSIX paths cannot be used on Windows")
+    def test_mkrelative_posix(self):
         self.assertEqual(b'//asset.png', BlendPath.mkrelative(
             Path('/path/to/asset.png'),
             PurePosixPath('/path/to/bfile.blend'),
@@ -109,3 +113,140 @@ class BlendPathTest(unittest.TestCase):
             Path('/shallow/asset.png'),
             PurePosixPath('/path/to/very/very/very/very/very/deep/bfile.blend'),
         ))
+
+    @unittest.skipIf(platform.system() != 'Windows', "Windows paths cannot be used on POSIX")
+    def test_mkrelative_windows(self):
+        self.assertEqual(b'//asset.png', BlendPath.mkrelative(
+            Path('Q:/path/to/asset.png'),
+            PureWindowsPath('Q:/path/to/bfile.blend'),
+        ))
+        self.assertEqual(b'//to/asset.png', BlendPath.mkrelative(
+            Path('Q:/path/to/asset.png'),
+            PureWindowsPath('Q:/path/bfile.blend'),
+        ))
+        self.assertEqual(b'//../of/asset.png', BlendPath.mkrelative(
+            Path('Q:/path/of/asset.png'),
+            PureWindowsPath('Q:/path/to/bfile.blend'),
+        ))
+        self.assertEqual(b'//../../path/of/asset.png', BlendPath.mkrelative(
+            Path('Q:/path/of/asset.png'),
+            PureWindowsPath('Q:/some/weird/bfile.blend'),
+        ))
+        self.assertEqual(b'//very/very/very/very/very/deep/asset.png', BlendPath.mkrelative(
+            Path('Q:/path/to/very/very/very/very/very/deep/asset.png'),
+            PureWindowsPath('Q:/path/to/bfile.blend'),
+        ))
+        self.assertEqual(b'//../../../../../../../../shallow/asset.png', BlendPath.mkrelative(
+            Path('Q:/shallow/asset.png'),
+            PureWindowsPath('Q:/path/to/very/very/very/very/very/deep/bfile.blend'),
+        ))
+        self.assertEqual(b'D:/path/to/asset.png', BlendPath.mkrelative(
+            Path('D:/path/to/asset.png'),
+            PureWindowsPath('Q:/path/to/bfile.blend'),
+        ))
+
+
+class MakeAbsoluteTest(unittest.TestCase):
+    def test_relative(self):
+        my_dir = Path(__file__).absolute().parent
+        cwd = os.getcwd()
+        try:
+            os.chdir(my_dir)
+            self.assertEqual(my_dir / 'blendfiles/Cube.btx',
+                             make_absolute(Path('blendfiles/Cube.btx')))
+        except Exception:
+            os.chdir(cwd)
+            raise
+
+    @unittest.skipIf(platform.system() != 'Windows', "This test uses drive letters")
+    def test_relative_drive(self):
+        cwd = os.getcwd()
+        my_drive = Path(f'{Path(cwd).drive}/')
+        self.assertEqual(my_drive / 'blendfiles/Cube.btx',
+                         make_absolute(Path('/blendfiles/Cube.btx')))
+
+    def test_drive_letters(self):
+        """PureWindowsPath should be accepted and work well on POSIX systems too."""
+        in_path = PureWindowsPath('R:/wrongroot/oops/../../path/to/a/file')
+        expect_path = Path('R:/path/to/a/file')
+        self.assertNotEqual(expect_path, in_path, 'pathlib should not automatically resolve ../')
+        self.assertEqual(expect_path, make_absolute(in_path))
+
+    @unittest.skipIf(platform.system() == 'Windows', "This test ignores drive letters")
+    def test_dotdot_dotdot_posix(self):
+        in_path = Path('/wrongroot/oops/../../path/to/a/file')
+        expect_path = Path('/path/to/a/file')
+        self.assertNotEqual(expect_path, in_path, 'pathlib should not automatically resolve ../')
+        self.assertEqual(expect_path, make_absolute(in_path))
+
+    @unittest.skipIf(platform.system() != 'Windows', "This test uses drive letters")
+    def test_dotdot_dotdot_windows(self):
+        in_path = Path('Q:/wrongroot/oops/../../path/to/a/file')
+        expect_path = Path('Q:/path/to/a/file')
+        self.assertNotEqual(expect_path, in_path, 'pathlib should not automatically resolve ../')
+        self.assertEqual(expect_path, make_absolute(in_path))
+
+    @unittest.skipIf(platform.system() == 'Windows', "This test ignores drive letters")
+    def test_way_too_many_dotdot_posix(self):
+        in_path = Path('/webroot/../../../../../etc/passwd')
+        expect_path = Path('/etc/passwd')
+        self.assertEqual(expect_path, make_absolute(in_path))
+
+    @unittest.skipIf(platform.system() != 'Windows', "This test uses drive letters")
+    def test_way_too_many_dotdot_windows(self):
+        in_path = Path('G:/webroot/../../../../../etc/passwd')
+        expect_path = Path('G:/etc/passwd')
+        self.assertEqual(expect_path, make_absolute(in_path))
+
+    @unittest.skipIf(platform.system() == 'Windows',
+                     "Symlinks on Windows require Administrator rights")
+    def test_symlinks(self):
+        with tempfile.TemporaryDirectory(suffix="-bat-symlink-test") as tmpdir_str:
+            tmpdir = Path(tmpdir_str)
+
+            orig_path = tmpdir / 'some_file.txt'
+            with orig_path.open('w') as outfile:
+                outfile.write('this file exists now')
+
+            symlink = tmpdir / 'subdir' / 'linked.txt'
+            symlink.parent.mkdir()
+            symlink.symlink_to(orig_path)
+
+            self.assertEqual(symlink, make_absolute(symlink), 'Symlinks should not be resolved')
+
+    @unittest.skipIf(platform.system() != 'Windows',
+                     "Drive letters mapped to network share can only be tested on Windows")
+    @unittest.skip('Mapped drive letter testing should be mocked, but that is hard to do')
+    def test_mapped_drive_letters(self):
+        pass
+
+    def test_path_types(self):
+        platorm_path = type(PurePath())
+        self.assertIsInstance(make_absolute(PureWindowsPath('/some/path')), platorm_path)
+        self.assertIsInstance(make_absolute(PurePosixPath('/some/path')), platorm_path)
+
+
+class StripRootTest(unittest.TestCase):
+    def test_windows_paths(self):
+        self.assertEqual(PurePosixPath(), strip_root(PureWindowsPath()))
+        self.assertEqual(
+            PurePosixPath('C/Program Files/Blender'),
+            strip_root(PureWindowsPath('C:/Program Files/Blender')))
+        self.assertEqual(
+            PurePosixPath('C/Program Files/Blender'),
+            strip_root(PureWindowsPath('C:\\Program Files\\Blender')))
+        self.assertEqual(
+            PurePosixPath('C/Program Files/Blender'),
+            strip_root(PureWindowsPath('C\\Program Files\\Blender')))
+
+    def test_posix_paths(self):
+        self.assertEqual(PurePosixPath(), strip_root(PurePosixPath()))
+        self.assertEqual(
+            PurePosixPath('C/path/to/blender'),
+            strip_root(PurePosixPath('C:/path/to/blender')))
+        self.assertEqual(
+            PurePosixPath('C/path/to/blender'),
+            strip_root(PurePosixPath('C/path/to/blender')))
+        self.assertEqual(
+            PurePosixPath('C/path/to/blender'),
+            strip_root(PurePosixPath('/C/path/to/blender')))